Move SpecialField and SpecialFieldUtil to logic
