+++ src/main/java/net/sf/jabref/gui/maintable/MainTableSelectionListener.java
@@ -36,12 +36,12 @@ import net.sf.jabref.gui.menus.RightClickMenu;
 import net.sf.jabref.gui.specialfields.SpecialFieldActions;
 import net.sf.jabref.gui.specialfields.SpecialFieldMenuAction;
 import net.sf.jabref.logic.l10n.Localization;
+import net.sf.jabref.logic.specialfields.SpecialFieldsUtils;
 import net.sf.jabref.logic.util.OS;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.entry.SpecialFieldValue;
 import net.sf.jabref.preferences.PreviewPreferences;
-import net.sf.jabref.specialfields.SpecialFieldsUtils;
 
 import ca.odell.glazedlists.EventList;
 import ca.odell.glazedlists.event.ListEvent;
