+++ src/main/java/net/sf/jabref/gui/specialfields/SpecialFieldUpdateListener.java
@@ -4,10 +4,10 @@ import javax.swing.SwingUtilities;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefGUI;
+import net.sf.jabref.logic.specialfields.SpecialFieldsUtils;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.entry.event.FieldChangedEvent;
-import net.sf.jabref.specialfields.SpecialFieldsUtils;
 
 import com.google.common.eventbus.Subscribe;
